se refactoriza nombre de package en pages
