+++ src/test/java/aut/testcreation/pages/rumboesbracamontequezadapino/FlashSalesPage.java
@@ -1,4 +1,4 @@
-package aut.testcreation.pages.rumboesabbaro;
+package aut.testcreation.pages.rumboesbracamontequezadapino;
 
 import framework.engine.selenium.SeleniumWrapper;
 import org.openqa.selenium.WebDriver;
