+++ src/test/java/aut/testcreation/pages/rumboesbracamontequezadapino/vuelosPages/VuelosBusquedaPage.java
@@ -1,4 +1,4 @@
-package aut.testcreation.pages.rumboesabbaro.vuelosPages;
+package aut.testcreation.pages.rumboesbracamontequezadapino.vuelosPages;
 
 import framework.engine.selenium.SeleniumWrapper;
 import org.openqa.selenium.By;
